pyproject.toml
[project]
name = "datamasters-mlops-case"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "mlflow==2.16.0",
    "scikit-learn==1.5.2",
    "psycopg2-binary==2.9.9",
    "boto3==1.35.0",
    "xgboost==2.1.1",
    "fastapi==0.110.0",
    "starlette==0.36.3",
    "httpx==0.27.0",
    "uvicorn==0.30.0",
    "pydantic==2.6.4",
    "python-multipart",
]


[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "pytest",
    "pytest-mock",
    "pytest-cov",
    "python-dotenv",
    "httpx==0.27.0",
    "requests",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
exclude = ["data","models"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
